Major mods to "XR Device Simulator" script
Made some major modifications to the "XRDeviceSimulator"plugin script.  In its normal functionality, the controllers are in a fixes starting position which is right next to the camera/headset when simulating.

Now, you can set the positions of the controllers dynamically from the inspector before you press play.  Look for "Layla Custom Fields - Controllers" in the inspector on the Device Simulator.

Fixed some weirdness with the camera when simulating as well. changed the "Camera Offset" var in the XR input to 0, changed the height of the camera transform to 0, and changed the height of the "Camera Offset" OBJECT to 2.5
Everything is working as expected right now.

Had to update the .gitignore for this as well to not ignore the script, since it's in Library (which usually gets ignored)