Added more tvs/consoles panels to the station
Made 4 game objects named "Screens/Consoles Test [num]". (be aware all of then are disabled)
The game object that holds all 4 of them is named "SC Holder".

With each of them only having two to determine if they would want to be used or not since the station itself is really small and limited on where to put each panels to not make it look clutterd/unique of its own!

The gameobject that holds all 4 of them is named "SC Holder"